backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai==1.12.0
pyautogen==0.2.12
pydantic==2.6.1 